+++ src/main/java/at/hannibal2/skyhanni/events/SackChangeEvent.kt
@@ -1,5 +1,5 @@
 package at.hannibal2.skyhanni.events
 
-import at.hannibal2.skyhanni.features.inventory.SackAPI
+import at.hannibal2.skyhanni.data.SackAPI
 
 class SackChangeEvent(val sackChanges: List<SackAPI.SackChange>) : LorenzEvent()
